[osboot] Ensure loaded images are page aligned
UEFI's AllocatePool allocates memory but only ensures it is 8
byte aligned. Due to memory layout and size of images, this
can result in an unaligned ramdisk buffer being passed to
the OS which causes a silent assertion in userboot.

This is solved by switching to AllocatePages which maintains
that buffers are always page aligned.

Change-Id: I78f5bb4481bd91581e6ba3c1eece47bae907ba95