[netifc] Correct SimpleNetworkProtocol handling
OpenInterface was designed for driver use and can deadlock with existing
drivers already owning/providing a handle. Switching to HandleProtocol
allows us to properly utilize a provided driver handle as intended. This
means that we can use the ASIX driver provided by PXE software on
devices whose rom contain it already (Acer12), but we can also use our
own provided driver as necessary for platforms that do not (NUC).

Change-Id: I0b70f66ff3a05f8bfe56468f271fce7a76439b97