[gigaboot] Delete gnu-efi and stop doing relocations at runtime
With a carefully hand-crafted set of compiler and linker flags, it is
possible to build EFI images that do not need any relocations (for
x86-64 at least). This does forbid static initialization of addresses
though, which is not too big of a problem. The build will fail if an
image has relocations.

The entry point to our EFI images is now efi_main, but requires the
MS x64 calling convention. Prefix with EFIABI to use this convention.

Change-Id: I0e6fb38d6eb67aa79a9e5e5f3a1d0063e151355f